Move logger adapter implementation and formatters to the temp project.
